==> scripts/2.dw_survey_analysis.Rmd <==
---
title: "Untitled"
author: "abbey yatsko"
date: "11/4/2021"
output: html_document
---

read in data and set workspace 
```{r}
# call this data 'cleaned_survey_14-21'
data <- read.csv("/Users/abbeyyatsko/Desktop/repos/serc_dc/data/final_dwsurvey/cleaned_survey_14-21.csv")
library(ggplot2)
library(ggpubr)
library(dplyr)
library(tidyverse)
```

#### 1. INITIAL PLOTS FOR DW DISTRIBUTION ----
we want to know for each survey, what is the distribution of decay classes, and what form are DW pieces in (log or snag)?
cleaning data for breakdown status to agree across 3 surveys: 
```{r}
# fix the breakdown status for 2014 - change it from the werid language to just log and snag 
data <- data %>%
  mutate(BDS.2014_new = case_when(
    BDS.2014 == "lying broken" ~ "log", 
    BDS.2014 == "standing dead" ~ "snag",
    BDS.2014 == "standing broken" ~ "snag",
    BDS.2014 == "lying whole" ~ "log", 
    BDS.2014 == "lying cut" ~ "log",
    BDS.2014 == "standing cut" ~ "snag"
    ))

# fix the breakdown status for 2017
data <- data %>%
  mutate(BDS.2017_new = case_when(
    BDS.2017 == "Log" ~ "log", 
    BDS.2017 == "Snag" ~ "snag",
    BDS.2017 == "Stump" ~ "snag",
    BDS.2017 == "Consumed" ~ "consumed", 
    BDS.2017 == "NF" ~ "NF",
    ))

# fix the breakdown status for 2021 (one 'snag' has an extra space behind it)
data <- data %>%
  mutate(BDS.2021_new = case_when(
    BDS.2021 == "Snag " ~ "snag", 
    BDS.2021 == "Snag" ~ "snag",
    BDS.2021 == "Log" ~ "log"
    ))
```

Plot 1: decay class distribution by breakdown status:
create plots by survey year that show decay class and then the proportion of log / snag things that make up the count

```{r}

dc2014 <- ggplot(data = subset(data, !is.na(DC.2014)), aes(x=DC.2014, fill=BDS.2014_new)) + 
  geom_bar(position="stack", stat="count")+
  scale_fill_manual( breaks = c("log","snag"),
                    values = c("darkgreen", "orange", "red", "blue"))

dc2017 <- ggplot(data = subset(data[which(data$DC.2017 < 6),], !is.na(DC.2017)), aes(x=DC.2017, fill=BDS.2017)) + 
  geom_bar(position="stack", stat="count")+
  scale_fill_manual( breaks = c("Log", "Snag", "Consumed"),
                    values = c("darkgreen", "orange", "red", "blue"))

dc2021 <- ggplot(data = subset(data[which(data$DC.2021 < 6),], !is.na(DC.2021)), aes(x=DC.2021, fill=BDS.2021_new)) + 
  geom_bar(position="stack", stat="count")+
  scale_fill_manual( breaks = c("log", "snag"),
                    values = c("darkgreen", "orange", "red"))

ggarrange(dc2014, dc2017, dc2021, nrow = 1)

```

Plot 2: species distribution of deadwood in forestgeo plot
create plot that shows species distribution, which species are most common 

```{r}
speciesdist <- ggplot(data = subset(data, !is.na(SPCODE)), aes(x=SPCODE, fill=BDS.2021_new)) + 
  # include display of BDS for most recent survey (fill as breakdown status)
  theme_classic() +
  geom_bar(position="stack", stat="count")+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  scale_y_continuous(expand = c(0, 0), limits = c(0, 300)) 

```

#### 2. GETTING SURVEYS TO TALK ----

two problems: 
> 2014 is in three-tiers, while 2017/2021 is in 5 
> we could also split out the 2014 data into 5 tiers, but we need more information on how to do this 

this is code for how collapsing the 2017 and 2021 data could work (everythign would be in 3 tiers)
```{r}
# QUESTION: how to break up 1 into 1's and 2's as well as 2 into 3's and 4's? 
# when building transition matrix, need DC to be equivalent across the three surveys 
# currently, it is not like that! 

# this is the code for going the other direction: now we are splitting up all decay classes into 3 tier 
# convert 2017 and 2021 decay class data to the 3 tiered scale based on the above criteria: 

  # 1 = 1,2
  # 2 = 3,4
  # 3 = 5

# above we also created a DC 6 for those pieces that 'moved out of the system' aka were not able to be found or were listed as consumed. 
# add in a 4th decay class which translates to this 'consumed' phase 

  # 4 = 6 (consumed / left the system)

# in 'data' - create 2 new columns 'DC.2017_3' and 'DC.2021_3'
# not changing DC.2014 because it is already in terms of 3 tier decay class 

# for 2017 survey
data$DC.2017_3 <- ifelse((data$DC.2017 == "1") | (data$DC.2017 == "2"), "1",
                       ifelse((data$DC.2017 == "3") | (data$DC.2017 == "4"), "2",
                              ifelse((data$DC.2017 == "6"), NA, # here is where the 'left system' decay class is added 
                              ifelse((data$DC.2017 == "5"), "3", NA))))

unique(data$DC.2017_3)
str(data$DC.2017_3)
data$DC.2017_3 <- as.integer(data$DC.2017_3)

# plot this 
ggplot(data = subset(data, !is.na(DC.2017_3)), aes(x=DC.2017_3, fill=BDS.2017)) + 
  geom_bar(position="stack", stat="count")+
  scale_fill_manual( breaks = c("Log", "Snag", "Consumed"),
                    values = c("darkgreen", "orange", "red", "blue"))

# most of decay class 4 is consumed for BDS - this makes sense because it is what the class represents 
# there are many 1's and 2's, hardly any 3's 
# this makes sense if you run counts for the initial DC.2017 dataset number of 5's: 
# > sum(data$DC.2017 == "5", na.rm = TRUE)
# 6
# > sum(data$DC.2017 == "4", na.rm = TRUE)
# 133

# whether its categorized as a 5 or a 3, this decay class is going to have fewer members it seems 

# for 2021 survey
data$DC.2021_3 <- ifelse((data$DC.2021 == "1") | (data$DC.2021 == "2"), "1",
                       ifelse((data$DC.2021 == "3") | (data$DC.2021 == "4"), "2",
                            ifelse((data$DC.2021 == "6"), NA, # here is where the 'left system' decay class is added 
                              ifelse((data$DC.2021 == "5"), "3", NA))))

unique(data$DC.2021_3)
data$DC.2021_3 <- as.integer(data$DC.2021_3)

data[c("DC.2021", "DC.2021_3")]

# plot this 
ggplot(data = subset(data, !is.na(DC.2021_3)), aes(x=DC.2021_3, fill=BDS.2021)) + 
  geom_bar(position="stack", stat="count")+
  scale_fill_manual( breaks = c("Log", "Snag", "Consumed"),
                    values = c("darkgreen", "orange", "red", "blue", "black", "brown")) # weird error with number of things on the scale - added black and brown but it doesnt actually change the fig... 

# in this survey there are many more DC 3's. almost 10x as many
# i guess this indicates that 2017 survey was biased to a less decayed pool, but by 2021 some of those pieces have shifted? 

# there are also a fair number of pieces that have left the system (indicated by 4's) 

############################################################################################################

# FOR THE NOTES: ALTERNATE WAY of converting decay class - skip for now
# in new.df - create 2 new columns 'DC.2017_3' and 'DC.2021_3'
# 
# new.df <- new.df %>%
#   mutate(DC.2017_3 = case_when(
#     DC.2017 == "1" ~ "1", 
#     DC.2017 == "2" ~ "2",
#     DC.2017 == "3" ~ "2",
#     DC.2017 == "4" ~ "3", 
#     DC.2017 == "5" ~ "3",
#     ))
# select(new.df, DC.2017, DC.2017_3)
# 
# new.df <- new.df %>%
#   mutate(DC.2021_3 = case_when(
#     DC.2021 == "1" ~ "1", 
#     DC.2021 == "2" ~ "2",
#     DC.2021 == "3" ~ "2",
#     DC.2021 == "4" ~ "3", 
#     DC.2021 == "5" ~ "3",
#     ))
# select(new.df, PIECETAG, DC.2021, DC.2021_3)
# 
# tm_data <- select (new.df, PIECETAG, DC.2014, DC.2017_3, DC.2021_3)

```

now everything is in decay class scale 1-3! time to move on and compare

the next question is, what pieces can be tracked through time completely? this means that they were recorded in all 3 surveys 
samples that are noted 3 times give us the most complete record of the deadwood decomposition process 
```{r}
# dataframe of complete survey data for DW sample across all 3 survey periods 
a <- data %>% filter(!is.na(DC.2014), !is.na(DC.2017_3), !is.na(DC.2021_3))

# 2014 samples that are a 3 and then are not found in 2017 or in 2021
b <-  data %>% filter(!is.na(DC.2014), is.na(DC.2017_3), is.na(DC.2021_3))

# the question here is, do we assume that these pieces have moved 'out of the system'? 
# there is a mix of 1-3, maybe for 3's the assumption is that they moved out?

```

#### 3. EXPORT data to FINAL_DATA ----

```{r}
# call this data 'cleaned_survey_14-21'
write.csv(data, "/Users/abbeyyatsko/Desktop/repos/serc_dc/data/final_dwsurvey/transition_survey_14-21.csv")
```
