==> scripts/3.dc_transition_probabilities.Rmd <==
---
title: "decay probabilities"
author: "abbey yatsko"
date: "4/19/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### 1. READ IN DATA ----
```{r}
data <- read.csv("/Users/abbeyyatsko/Desktop/repos/serc_dc/data/final_dwsurvey/transition_survey_14-21.csv")
```

#### 2. DECAY CLASS PROBABILITY MATRIX FROM DATA (non-annualized) ----
main question: what are the probabilities of transition between the different states (1-3 decay classes) across three time points (2014, 2017, 2021)?

4 code chunks in this section: 
  * 2017-2021 (5 tier)
  * 2017-2021 (3 tier)
  * 2014-2017 (3 tier)
  * 2014-2021 (3 tier)
generating 4 unique probability matrices, which will need to be annualized in the next step

this section is just for transitioning from 2017 - 2021, but using 5 TIERS
```{r}
# create an empty 5x5 matrix
decay.mat1 <- matrix(0, nrow = 5, ncol = 5)

# instructions for how to fill in the matrix: 2017 is column, 2021 is row
for(i in 1:nrow(data)) {
	decay.mat1[data$DC.2021[i], data$DC.2017[i]] <- decay.mat1[data$DC.2021[i], data$DC.2017[i]] + 1
}

# decay.mat shows the absolute numbers of pieces that were moving between decay stages from 2017-2021
decay.mat1
# sum the columns to see how many 'came out the other end' of the transition
colSums(decay.mat1)
# table of data$DC.2017 shows initial inputs from 2017
table(data$DC.2017)
# there will be differences between table(data$DC.2017) and colSums(decay.mat1) - this represents pieces that did not survive 

# percent probability of going from a particular decay state in 2017 to the next state in 2021 
decay.prob1 <- sweep(decay.mat1,2,table(data$DC.2017),`/`) 
# need to be dividing by the table(data$DC.2017) counts rather than colSums(decay.mat1)
decay.prob1

# to check and make sure that survival is not 100% (which it should not be) - colSums should be < 1.0 
colSums(decay.prob1)

# checks out! also makes sense because values get smaller and smaller. 'leakage' at the 1-2 decay stages should be low - this implies that most 1's and 2's are moving through to other decay classes further along, but not disappearing all together. for the 4's and 5's, 'survival' is lower because we expect the probability of 'recapture' to be lower. essentially, pieces at 4/5 are more likely to be moving out of the system, i.e. going undetected in further surveys.
```

this section is just for transitioning from 2017 - 2021, but using 3 TIERS
```{r}
# create an empty 3x3 matrix
decay.mat2 <- matrix(0, nrow = 3, ncol = 3)

# instructions for how to fill in the matrix: 2017 is column, 2021 is row
for(i in 1:nrow(data)) {
	decay.mat2[data$DC.2021_3[i], data$DC.2017_3[i]] <- decay.mat2[data$DC.2021_3[i], data$DC.2017_3[i]] + 1
}

# percent probability of going from a particular decay state in 2017 to the next state in 2021 
decay.prob2 <- sweep(decay.mat2,2,table(data$DC.2017_3),`/`)
decay.prob2
colSums(decay.prob2)
```

this section is just for transitioning from 2014 - 2017, but using 3 TIERS
```{r}
# create an empty 3x3 matrix

decay.mat3 <- matrix(0, nrow = 3, ncol = 3)

# instructions for how to fill in the matrix: 2014 is column, 2017 is row
for(i in 1:nrow(data)) {
	decay.mat3[data$DC.2017_3[i], data$DC.2014[i]] <- decay.mat3[data$DC.2017_3[i], data$DC.2014[i]] + 1
}

# percent probability of going from a particular decay state in 2014 to the next state in 2017 
decay.prob3 <- sweep(decay.mat3,2,table(data$DC.2014),`/`)
decay.prob3
colSums(decay.prob3)
```

this section is just for transitioning from 2014 - 2021, but using 3 TIERS
```{r}
# create an empty 3x3 matrix

decay.mat4 <- matrix(0, nrow = 3, ncol = 3)

# instructions for how to fill in the matrix: 2014 is column, 2017 is row
for(i in 1:nrow(data)) {
	decay.mat4[data$DC.2021_3[i], data$DC.2014[i]] <- decay.mat4[data$DC.2021_3[i], data$DC.2014[i]] + 1
}

# percent probability of going from a particular decay state in 2014 to the next state in 2017 
decay.prob4 <- sweep(decay.mat4,2,table(data$DC.2014),`/`)
decay.prob4
colSums(decay.prob4)
```

interesting things going on with decay.prob3 and decay.prob4: the probability of a 3 reverting to a 2 from 2014 onward is higher than a 3 staying as a 3. this could be an artifact of converting between decay class scales.

need to address the converstion of a 3 tiered DC scale to that of 5 tiers. currently going under the assumption that: 
   1 = 1,2
   2 = 3,4
   3 = 5
   
rather than making this assumption, it would be better to have a statistically informed converstion of 3 tiered DC scale to 5 tiers

#### 3. FIXING THE 3 TIER / 5 TIER PROBLEM, CONSIDERING DIAMETERS ----

first look at what the problem is: how are pieces moving between 2014 and 2017 BEFORE converting decay class tiers to the same metric
```{r}
# create an empty 5x3 matrix
decay.mat_test <- matrix(0, nrow = 5, ncol = 3)

# instructions for how to fill in the matrix: 2014 is column, 2017 is row
for(i in 1:nrow(data)) {
  decay.mat_test[data$DC.2017[i], data$DC.2014[i]] <- decay.mat_test[data$DC.2017[i], data$DC.2014[i]] + 1
}

# percent probability of going from a particular decay state in 2014 to the next state in 2017 
decay.prob_test <- sweep(decay.mat_test,2,table(data$DC.2014),`/`)
decay.prob_test
colSums(decay.prob_test)

# why is decay class 2 disappearing? DC2 is smaller size...
tapply(data$DBH.2014, data$DC.2014, mean, na.rm=TRUE)

# conclude that it is important to include dbh within the decay transition models 
# smaller pieces are going to leave the system much faster than large pieces 
# 'leakage' is therefore biased for smaller diameter stems 
```

need to define a statistical way to convert from 3 tier to 5 tier (which is informed based on how things are actually moving between DC of different scales)

things that matter: 
  1. background rate of decomposition through time ('natural progression of decay')
  2. transition rate to new decay classes (taking into account different scales)
  3. effect of diameter on transition rate (both between DC and out of the system)

```{r}
# of all the things that didnt decay out of the system 2014-2017, where were they reassigned? 
# probabilities of being within each decay class
test <- apply(decay.prob_test, 2, function(x) x/sum(x))

# should all sum to 1 (now that we are excluding pieces that were 'lost to the system')
colSums(test)

# this should be the basis for a weighted average for reclssification between 3 tier to 5 tier scale
# i.e. a DC 1/3 has a 10% chance to go to a 1/5, a 57% chance to go to a 2/5, a 22% change to go to a 3/5, and a 11% chance to go to a 4/5. there is a 0% chance that a 1/3 will turn into a 5/5
```

now, see how diameter is distributed throughout the surveys
it is important to include diameter into the transition models as small things will be going quicker than larger things (contribute more to leakage)

```{r}
# distribution of stem DBH by survey
tapply(data$DBH.2014, data$DC.2014, summary, na.rm=TRUE)
tapply(data$DBH.2017, data$DC.2017, summary, na.rm=TRUE)
tapply(data$DBH.2021, data$DC.2021, summary, na.rm=TRUE)

# you can see the changes in protocol based on survey year - smaller things were measured in 2014, cutoff was definitely >10cm in 2017, some smaller than 10cm things made it in to the 2021 survey
```

get original DBH for all pieces - if we can get original DBH, then we have at least some starting information as to how big things were at the beginning of their decay process

```{r}
# make new dataframe DBH.df 
DBH.df <- data.frame(DBH.2014 = data$DBH.2014, DBH.2017 = data$DBH.2017, DBH.2021 = data$DBH.2021, DBH.x = data$DBH.x, DBH.y = data$DBH.y)
# figure out what exactly DBH.x and DBH.y refer to? 

# take the maximum value reported from DBH measured in 2014, 2017, and 2021, create new list of values "DBH.org"
DBH.org <- apply(DBH.df, 1, max, na.rm=TRUE)
head(DBH.df, 30)
head(DBH.org, 30)

# replace -Inf with NA (meaning that in 2014, 2017, and 2021 there is no recorded DBH in the deadwood survey dataset)
DBH.org[DBH.org==-Inf] <- NA

# full distribution of max size classes across the surveys (note the number of NAs - these will need to be filled in)
summary(DBH.org)

# compare dimensions to add DBH.org back into the working df 'data'
length(DBH.org)
dim(data)
data$DBH.org <- DBH.org

# look at the things that have NA in data$DBH.org, add in living DBH data
# DO THIS 
fixNAs <- data[is.na(data$DBH.org),]

# there are 57 pieces that need a DBH to be assigned to them - most are A pieces, but some are logs that have multiple segments (through E or F)

# does the piecetag show up in the main df? 

# if so, match the dbh from "data"

# 
```

#### 4. ANNUALIZE DECAY CLASS PROBABILITY MATRIX ----
since multiple years pass between each deadwood survey, the probabilities in the decay transition matrices need to be 'annualized' 

* 2014-2017: 3 years 
* 2017-2021: 4 years 
* 2014-2021: 7 years

the 'survival' also has to be taken into account and annualized - see code for this. it is necessary to make sure that we retain the same rates of 'leakage' where some pieces are not being re-found all together. 

5 code chunks in this section: 
  * proof of concept 
  * annualize matrix for 2017-2021 (5 tier, 4 years)
  * annualize matrix for 2017-2021 (3 tier, 4 years)
  * annualize matrix for 2014-2017 (3 tier, 3 years)
  * annualize matrix for 2014-2021 (3 tier, 7 years)
  
proof of concept for annualizing transition matrix probabilities 
```{r}
yrs <- 5 # number of years between censuses

A <- matrix(c(0.5, 0.25, 0, 0.5), ncol = 2) # Transition probabilities from column state at time t to row state at time t + yrs

Asurv <- colSums(A) ^ (1 / yrs) # Annualized survival prob
A_y <- A ^ (1 / yrs) # Matrix of annualized transitions

Ayr <- matrix(NA, ncol = 2, nrow = 2) # new annualized matrix

for(i in 1:dim(Ayr)[1]) {
  Ayr[, i] <- (A_y[, i] / colSums(A_y)[i]) * Asurv[i] 
# We now standardize not by 1, but by the annualized survival rate, so that the stages (columns) still 'leak' individuals at the adjusted annual rate 
}
```

annualize matrix for 2017-2021 (5 tier, 4 years)
```{r}
# un-annualized matrix 
decay.prob1

# number of years between 2017-2021 survey
yrs17to21 <- 4

# annualized survival probability
decay.prob1surv <- colSums(decay.prob1) ^ (1 / yrs17to21) 

# matrix of annualized transitions
decay.prob1_y <- decay.prob1 ^ (1 / yrs17to21) 

# new annualized matrix to fill in 
decay.prob1yr <- matrix(NA, ncol = 5, nrow = 5) 

# now standardize not by 1, but by the annualized survival rate, so that the stages (columns) still 'leak' individuals at the adjusted annual rate 
for(i in 1:dim(decay.prob1yr)[1]) {
  decay.prob1yr[, i] <- (decay.prob1_y[, i] / colSums(decay.prob1_y)[i]) * decay.prob1surv[i]
}

decay.prob1yr

# 'decayprob1yr' should be used now for time series plot 
# how to check if this works: (colSums(decay.prob1yr)) ^ 4 = colSums(decay.prob1)
```

annualize matrix for 2017-2021 (3 tier, 4 years)
```{r}
# un-annualized matrix 
decay.prob2

# annualized survival probability
decay.prob2surv <- colSums(decay.prob2) ^ (1 / yrs17to21) 

# matrix of annualized transitions
decay.prob2_y <- decay.prob2 ^ (1 / yrs17to21) 

# new annualized matrix to fill in 
decay.prob2yr <- matrix(NA, ncol = 3, nrow = 3) 

# now standardize not by 1, but by the annualized survival rate, so that the stages (columns) still 'leak' individuals at the adjusted annual rate 
for(i in 1:dim(decay.prob2yr)[1]) {
  decay.prob2yr[, i] <- (decay.prob2_y[, i] / colSums(decay.prob2_y)[i]) * decay.prob2surv[i]
}

decay.prob2yr
```

annualize matrix for 2014-2017 (3 tier, 3 years)
```{r}
# un-annualized matrix 
decay.prob3

# number of years between 2014-2017 survey
yrs14to17 <- 3

# annualized survival probability
decay.prob3surv <- colSums(decay.prob3) ^ (1 / yrs14to17) 

# matrix of annualized transitions
decay.prob3_y <- decay.prob3 ^ (1 / yrs14to17) 

# new annualized matrix to fill in 
decay.prob3yr <- matrix(NA, ncol = 3, nrow = 3) 

# now standardize not by 1, but by the annualized survival rate, so that the stages (columns) still 'leak' individuals at the adjusted annual rate 
for(i in 1:dim(decay.prob3yr)[1]) {
  decay.prob3yr[, i] <- (decay.prob3_y[, i] / colSums(decay.prob3_y)[i]) * decay.prob3surv[i]
}

decay.prob3yr
```

annualize matrix for 2014-2021 (3 tier, 7 years)
```{r}
# un-annualized matrix 
decay.prob4

# number of years between 2014-2017 survey
yrs14to21 <- 7

# annualized survival probability
decay.prob4surv <- colSums(decay.prob4) ^ (1 / yrs14to21) 

# matrix of annualized transitions
decay.prob4_y <- decay.prob4 ^ (1 / yrs14to21) 

# new annualized matrix to fill in 
decay.prob4yr <- matrix(NA, ncol = 3, nrow = 3) 

# now standardize not by 1, but by the annualized survival rate, so that the stages (columns) still 'leak' individuals at the adjusted annual rate 
for(i in 1:dim(decay.prob4yr)[1]) {
  decay.prob4yr[, i] <- (decay.prob4_y[, i] / colSums(decay.prob4_y)[i]) * decay.prob4surv[i]
}

decay.prob4yr
```

#### 5. CREATE DECAY CLASS TRANSITION SIMULATIONS (no pulse) ----
now that the transition matrices are annualized, it is time to start with visualizations and time series plots to see how the deadwood pool changes over time 

X code chunks in this section: 
  * simulated matrix example as proof of concept
  * DW simulation for annualized matrix (2017-2021)
  * DW simulation for annualized matrix (2017-2021)
  * DW simulation for annualized matrix (2014-2017)
  * DW simulation for annualized matrix (2014-2021)

first start with a simulated matrix example: 
```{r}
# set up an empty matrix
A <- matrix(0, 3, 3)

# populate matrix with different 'decay transition probabilities'
A[1, 1] <- 0.6
A[2, 1] <- 0.3
A[3, 1] <- 0.1
A[1, 2] <- 0.05
A[2, 2] <- 0.6
A[3, 2] <- 0.35
A[1, 3] <- 0.01
A[2, 3] <- 0.1
A[3, 3] <- 0.89

# see what the matrix looks like 
A

# pick a timeframe for the simulation to run over
yrs <- 25

# create a matrix (N_tot) for running the simulation, where the columns are each year of 
# the simulation and the rows are the proportion of pieces in each decay class, by year
N_tot <- matrix(0, ncol = yrs, nrow = 3)

# in the first year, put in 30 pieces of decay class 1 wood into the system 
N_tot[, 1] <- matrix(c(30, 0, 0), ncol = 1)

# start the simulation in the second year. use the inputs from the first year 
# (30 pieces of decay class 1 wood) and run them against the probability matrix 'A'
for(i in 2:yrs) {
	N_tot[, i] <- N_tot[, i - 1] %*% A
}

# look at the transition matrix simulated out to X years 
N_tot 

# display time series matrix simulation (N_tot) as a time series plot
ts.plot(t(N_tot), col = c(1, 2, 3))

# color legend: 
# DC1 = black
# DC2 = red 
# DC3 = green

# play around with different values to see consequences of different decay class pool strengths 
```

DW simulation for annualized matrix (2017-2021, 5 tier)
```{r}
decay.prob1yr
# yrs <- 25

# create a matrix (N_tot) for running the simulation out to the value of 'yrs'
N_tot1 <- matrix(0, ncol = yrs, nrow = 5)

# input of 30 pieces of decay class 1 wood into the system at t1
N_tot1[, 1] <- matrix(c(30, 0, 0, 0, 0), ncol = 1)

# start the simulation in the second year
# use the inputs from the first year (30 pieces of decay class 1 wood) and run them against the annualized probability matrix 'decay.prob1yr'
for(i in 2:yrs) {
	N_tot1[, i] <- N_tot1[, i - 1] %*% decay.prob1yr
}

# look at the transition matrix simulated out to X years 
N_tot1

# display time series matrix simulation (N_tot) as a time series plot
ts.plot(t(N_tot1), col = c(1, 2, 3, 4, 5))
```

DW simulation for annualized matrix (2017-2021, 3 tier)
```{r}
decay.prob2yr
N_tot2 <- matrix(0, ncol = yrs, nrow = 3)

N_tot2[, 1] <- matrix(c(30, 0, 0), ncol = 1)

for(i in 2:yrs) {
	N_tot2[, i] <- N_tot2[, i - 1] %*% decay.prob2yr
}

N_tot2

ts.plot(t(N_tot2), col = c(1, 2, 3))
```

DW simulation for annualized matrix (2014-2017)
```{r}
decay.prob3yr
N_tot3 <- matrix(0, ncol = yrs, nrow = 3)

N_tot3[, 1] <- matrix(c(30, 0, 0), ncol = 1)

for(i in 2:yrs) {
	N_tot3[, i] <- N_tot2[, i - 1] %*% decay.prob3yr
}

N_tot3

ts.plot(t(N_tot3), col = c(1, 2, 3))
```

DW simulation for annualized matrix (2014-2021)
```{r}
decay.prob4yr
N_tot4 <- matrix(0, ncol = yrs, nrow = 3)

N_tot4[, 1] <- matrix(c(30, 0, 0), ncol = 1)

for(i in 2:yrs) {
	N_tot4[, i] <- N_tot4[, i - 1] %*% decay.prob4yr
}

N_tot4

ts.plot(t(N_tot4), col = c(1, 2, 3))
```

#### 6. CREATE DECAY CLASS TRANSITION SIMULATIONS (with pulse) ----
the above examples show how one pulse at t1 affects decay class pools throughout time, but without annual additions of new deadwood. 

next goal is to make simulations where repeated influxes are made at each time point

TRIAL: add pulses for decay.prob1yr transition matrix (2017-2021 with 5 tiers) to t4

NOTE: need to figure out how to make this into a loop - ask sean 
```{r}
# starting annualized transition matrix (representing 2017-2021 on 5 tiers)
decay.prob1yr
#           [,1]       [,2]      [,3]      [,4]      [,5]
# [1,] 0.20549134 0.09763219 0.0000000 0.0000000 0.0000000
# [2,] 0.31729342 0.24705236 0.1778925 0.1629008 0.0000000
# [3,] 0.22008484 0.26342804 0.2867369 0.2332037 0.0000000
# [4,] 0.15404334 0.22458593 0.2535208 0.2821524 0.0000000
# [5,] 0.09842493 0.13807277 0.2316419 0.2505937 0.7598357

# recreate this matrix for sean 
# decay.prob1yr <- matrix(c(0.20549134, 0.31729342, 0.22008484, 0.15404334, 0.09842493, 
#                           0.09763219, 0.24705236, 0.26342804, 0.22458593, 0.13807277, 
#                           0.0000000, 0.1778925, 0.2867369, 0.2535208, 0.2316419, 
#                           0.0000000, 0.1629008, 0.2332037, 0.2821524, 0.2505937, 
#                           0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.7598357),
#                         nrow = 5, ncol = 5)

# t0 Start
# start out with 50 DC = 1 at t0
N <- matrix(0, ncol = 1, nrow = 5)
N[, 1] <- matrix(c(50, 0, 0, 0, 0), ncol = 1)
N

# save t0 deadwood sum (representing only total in DC1 since this is the only thing in the system currently)
t0 <- matrix(rowSums(N), ncol = 1)

# move t0 through matrix
N <- (sweep(decay.prob1yr, MARGIN=2, t0, `*`)) 

# check: sum of N[, 1] should be ~50
sum(N[, 1])
# why is it not exactly 50?? 

# t1: add in 10 as the DC = 1 pulse
N[1,1] <- N[1,1] + 10

# save t1 deadwood sum 
t1 <- matrix(rowSums(N), ncol = 1)

# now, input t1 into the transition matrix in t1
N <- (sweep(decay.prob1yr, MARGIN=2, t1, `*`)) 

# add in 10 as the DC = 1 pulse
N[1,1] <- N[1,1] + 10

# save t2 deadwood sum 
t2 <- matrix(rowSums(N), ncol = 1)

# input t2 into the transition matrix
N <- (sweep(decay.prob1yr, MARGIN=2, t2, `*`)) 

# add in 10 as the DC = 1 pulse
N[1,1] <- N[1,1] + 10

# save t3 deadwood sum 
t3 <- matrix(rowSums(N), ncol = 1)

# input t3 into the transition matrix
N <- (sweep(decay.prob1yr, MARGIN=2, t3, `*`)) 

# add in 10 as the DC = 1 pulse
N[1,1] <- N[1,1] + 10

# save t4 deadwood sum 
t4 <- matrix(rowSums(N), ncol = 1)

# combine t0 - t4 into a dataframe to graph the t0 - t4 'time series'
df <- data.frame(t0, t1, t2, t3, t4)

# plot the 'timeseries' 
matplot(t(df),type="l")
```

GOAL is to have this in some kind of loop that can be simulated out to X years!

i made progress 
did I ? 

DRAFT STUFF HERE AND BELOW:
##################################################################
##################################################################
##################################################################

turn back to a simulated matrix, 'A', and build in new elements (see notes with sean)
```{r}
# matrix, but add confidence intervals for prediction values 
A[1, 1] <- rnorm(1, 0.6, 0.1)
A[2, 1] <- rnorm(1, 0.3, 0.15)
A[3, 1] <- rnorm(1, 0.1, 0.2)
A[1, 2] <- rnorm(1, 0.05, 0.1)
A[2, 2] <- rnorm(1, 0.6, 0.05)
A[3, 2] <- rnorm(1, 0.35, 0.11)
A[1, 3] <- rnorm(1, 0.01, 0.01)
A[2, 3] <- rnorm(1, 0.1, 0.1)
A[3, 3] <- rnorm(1, 0.89, 0.15)
A

# timeframe and populating matrix setup 
yrs <- 50
N_tot <- matrix(0, ncol = yrs, nrow = 3)
N_tot[, 1] <- matrix(c(30, 0, 0), ncol = 1)

for(i in 2:yrs) {
	N_tot[, i] <- N_tot[, i - 1] %*% A
}

# look at the transition matrix simulated out to X years 
N_tot 

# display time series matrix simulation (N_tot) as a time series plot
ts.plot(t(N_tot), col = c("black", "red", "green"))

# use "plot" and "lines" 
# cycle through and plot every line - creates distribution 
# add = true for ts.plot (check)

################################################################################

# UNFINISHED 

# NEXT TRY
# two loops where the first now feeds into the second (old loop)
# in the first loop you are sampling the larger dataset in order to make multiuple 'A' matrices 
#for(j in 1:samp) {  

# then add in the j randomly sampled A matrices into the loop
# for(i in 2:yrs) {
	N[, i, j] <- N_tot[, i - 1] %*% A
#}
#}
	
ts.plot(t(N_tot), col = c(1, 2, 3))
samp <- 1000
N <- array(0, dim = c(3, yrs, samp))

# use 'apply()' to get means and quantiles from the N array.

# hint: lines()
```

What about comparisons of decay class transition matrix between snags and logs? 
```{r}

# create an empty 2x2 matrix
decay.mat <- matrix(0, nrow = 2, ncol = 2)

# instructions for how to fill in the matrix: 2017 is column, 2021 is row
for(i in 1:nrow(data)) {
	decay.mat[data$BDS.2021_new[i], data$BDS.2017_new[i]] <- decay.mat[data$BDS.2021_new[i], data$BDS.2017_new[i]] + 1
}

# percent probability of going from a particular decay state in 2017 to the next state in 2021 
decay.prob1 <- sweep(decay.mat,2,colSums(decay.mat),`/`)
decay.prob1


```
